moved checks to single file
